[FEAT]: Symbol 에 따른 UI 변경 코드 추가
